+++ src/_packages/core.index.ts
@@ -1,5 +1,4 @@
-import { PromptTemplatePipelineLibrary } from '../classes/PromptTemplatePipelineLibrary';
-import { PTBK_VERSION } from '../version';
+import { PromptbookLibrary } from '../classes/PromptbookLibrary';
 import { promptTemplatePipelineStringToJson } from '../conversion/promptTemplatePipelineStringToJson';
 import { validatePromptTemplatePipelineJson } from '../conversion/validatePromptTemplatePipelineJson';
 import { createPtpExecutor } from '../execution/createPtpExecutor';
@@ -8,9 +7,10 @@ import { CallbackInterfaceTools } from '../execution/plugins/user-interface-exec
 import { CallbackInterfaceToolsOptions } from '../execution/plugins/user-interface-execution-tools/callback/CallbackInterfaceToolsOptions';
 import { SimplePromptInterfaceTools } from '../execution/plugins/user-interface-execution-tools/simple-prompt/SimplePromptInterfaceTools';
 import { ExecutionTypes } from '../types/ExecutionTypes';
+import { PTBK_VERSION } from '../version';
 
 // @promptbook/core
-export { ExecutionTypes, PTBK_VERSION, PromptTemplatePipelineLibrary };
+export { ExecutionTypes, PTBK_VERSION, PromptbookLibrary };
 
 // @promptbook/simple-prompt
 export { SimplePromptInterfaceTools };